move (important) diagnostic tests out of sandbox
